feat: support non-6-decimal vault assets (e.g. WETH) with real oracle-normalized USD floors
The vault assumed a 6-decimal (USDC-style) deposit asset in several
places. Two distinct bug classes, fixed differently:

1. Self-derivable decimals bugs — BufferManager.minRebalanceAmount and
   FixedMaturityModule's funding overflow guard were hardcoded *1e6
   literals. Both now derive from the deployed asset's actual
   decimals() at the point they're set/checked.

2. Real USD-normalization bugs — RouterRebalanceGuard/ExecutionMemory
   have fields named "Usd" (minMoveUsd, baseGasCostUsd, ...) meant to
   be fixed-dollar floors, but they were compared directly against raw
   asset-unit plan values with no oracle conversion. New
   OracleValuationLib fetches a fresh oracle price (fail-closed on
   stale/unset) and converts these WAD-USD constants into the plan's
   asset-unit terms at evaluation time. ExecutionMemory's cost-tracking
   types are widened uint32/uint64 -> uint256 since WAD-scale ($*1e18)
   amounts no longer fit the old 6dp-sized (*1e6) types.

   RouterAllocationPolicy's internal "Usd" fields (totalMoveUsd, etc.)
   are intentionally left as raw asset units — every consumer only
   compares them in same-unit ratios/bps against tvl, so they don't
   need oracle conversion. Documented in place to make this explicit.

Also completes two previously dead per-vault override paths in
GlobalConfig (setVaultWithdrawalOverride, setVaultGovCaps) needed for
non-USDC vaults, since GlobalConfig's global defaults are shared
across vaults and can't be decimals-aware by construction.

Adds ~48 new tests across 6 files, including a paired 6dp-USDC/18dp-
WETH parity test proving the oracle conversion is correct (not just
non-reverting), and a full-stack integration test wiring real
GlobalConfig/BufferManager/Policy/Guard/ExecutionMemory against an
18-decimal asset end to end.

+++ src/interfaces/IExecutionMemory.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.28;
 /// @notice Bootstrap-safe cost model with observationCount thresholds and inactivity decay.
 interface IExecutionMemory {
     struct ExecutionRecord {
-        uint64 emaGasCost;
+        uint256 emaGasCost; // WAD-scaled USD; widened from uint64 to hold non-6dp-scale costs
         uint32 emaSlippageBps;
         uint32 failedCount;
         uint32 successCount;
@@ -37,7 +37,7 @@ interface IExecutionMemory {
 
     /// @notice Record lookup
     function records(address strategy) external view returns (
-        uint64 emaGasCost,
+        uint256 emaGasCost,
         uint32 emaSlippageBps,
         uint32 failedCount,
         uint32 successCount,
